Convert the edit file to OWL Functional Syntax.
It's 2025, there's no reason to keep using OBO as the
edit format. Besides, we already need some axioms that
cannot be represented in OBO (in fbbi-ext.ofn).

So we merge the existing fbbi-edit.obo file with the fbbi-ext.ofn
extension and convert the result into OFN, and we will use that as the
new edit file.